march 26th deploy

+++ src/controllers/videoController.js
@@ -1,4 +1,29 @@
 import Video from "../models/Video"
+import User from "../models/User"
+import {Storage} from "@google-cloud/storage";
+
+
+const storage = new Storage({
+    projectId: process.env.GCLOUD_PROJECT,
+});
+// https://stackoverflow.com/questions/72605933/google-cloud-storage-file-stream-into-fsreadstream-without-saving-to-file
+async function generateSignedUrl(){
+    const fileName = "uploads/videos/MyRecording.webm";
+    const options = {
+        version:'v2',
+        action:'read',
+        expires:Date.now() + 1000*60*60,
+    };
+    const [url] = await storage
+    .bucket(process.env.GCS_BUCKET)
+    .file(fileName)
+    .getSignedUrl(options);
+
+    console.log(`The signed url for ${fileName} is ${url}.`);
+
+    // generateSignedUrl().catch(console.error);
+    return url;
+}
 
 
 export const trend = async (req,res) => {
@@ -9,7 +34,7 @@ export const trend = async (req,res) => {
     // });
     // Method 2: promise
     try{
-        const videos = await Video.find({}).sort({createdAt:"asc"});
+        const videos = await Video.find({}).sort({createdAt:"asc"}).populate("owner");
 
         return res.render("home",{pageTitle:"Home", videos});
     } catch(error){
@@ -18,8 +43,15 @@ export const trend = async (req,res) => {
     }
 };
 export const watch = async (req,res) => {
+    // testing 
+    const url = await generateSignedUrl();
+
     const {id} = req.params;
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("owner");
+    
+    // need to change this. 
+    video.fileUrl = url;
+
     if(video === null){
         return res.status(404).render("404",{pageTitle:"Video not found"});
     }
@@ -28,16 +60,21 @@ export const watch = async (req,res) => {
 };
 export const getEdit = async (req,res) => {
     const {id} = req.params;
+    const {user:{_id}} = req.session;
     const video = await Video.findById(id);
     if(video === null){
         return res.status(404).render("404",{pageTitle:"Video not found"});
     }
 
+    if(String(video.owner) !== String(_id) ){
+        return res.status(403).redirect("/");
+    }
     return res.render("edit",{pageTitle:`Editt ${video.title}`,video});
 };
 
 export const postEdit = async (req,res) => {
     const {id} = req.params;
+    const {user:{_id},} = req.session;
     const { title , description, hashtags} = req.body;
     const video = await Video.exists({_id:id});
 
@@ -49,6 +86,10 @@ export const postEdit = async (req,res) => {
         description, 
         hashtags: Video.formatHashtags(hashtags),
     });
+    if(String(video.owner) !== String(_id) ){
+        return res.status(403).redirect("/");
+    }
+
     return res.redirect(`/videos/${id}`);
 };
 
@@ -57,13 +98,16 @@ export const getUpload = (req,res) =>{
 }
 
 export const postUpload = async (req,res)=>{
-    const {path:fileUrl} = req.file;
+    const {user:{_id}} = req.session;
+    const {video, thumb} = req.files;
     const {title,description, hashtags} = req.body;
     try{
-        await Video.create({
+        const newVideo = await Video.create({
             title: title,
             description : description, 
-            fileUrl,
+            fileUrl:video[0].path,
+            owner:_id,
+            thumbUrl:thumb[0].path,
             date: Date.now(),
             hashtags: Video.formatHashtags(hashtags),
             meta:{
@@ -71,6 +115,9 @@ export const postUpload = async (req,res)=>{
                 rating: 0,
             },
         });
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);
+        user.save();
         return res.redirect("/");
 
     } catch(error){
@@ -80,8 +127,19 @@ export const postUpload = async (req,res)=>{
 
 export const deleteVideo= async (req,res)=>{
     const {id} = req.params;
+    const {user:{_id},} = req.session;
+    const video = await Video.findById(id);
+    if(!video){
+        return res.status(404).render("404",{pageTitle:"Video not found."});
+    }
+    if(String(video.owner) !== String(_id) ){
+        return res.status(403).redirect("/");
+    }
+
     await Video.findByIdAndDelete(id);
     //delete video
+
+
     return res.redirect("/");
 
 }
@@ -92,4 +150,16 @@ export const search = (req,res) =>{
         //search
     }
     res.render("search",{pageTitle:"Search"});
+}
+
+export const registerView = async (req,res) =>{
+    const {id} = req.params;
+    const video = await Video.findById(id);
+
+    if(!video) {
+        return res.sendStatus(404);
+    }
+    video.meta.views = video.meta.views +1;
+    await video.save();
+    return res.sendStatus(200);
 }
